Added data pruning and passing that data to WEKA algorithms such as NaiveBayes.
Made AlgorithmsWEKA a superclass for ProcessData because it made sense, also probably removed some overhead and cleared up a small bit of code with this.
Changed NaiveBayes method to now take two datasets and instead of using "crossValidateModel" now it uses evaluateModel() method. Did this because we are required to compare two datasets and evaluate them both, instead of cross validating one dataset multiple times.

Changed (and added an additional) methods in FileHandler to protected static, so that they can be used to convert datasets to nominal without the need to create additional objects in the respective classes they're being called into.
Changed the "Run" button to now start the data pruning process. (Currently it doesn't start the NaiveBayes algorithm due to a bug, but that feature is planned)