Changed datasets from nominal to binary.
Currently the pruned dataset is not being used for testing purposes,
and the algorithm seems to be working nicely with any similar dataset with binary attributes.

Fixed the bug where data wouldn't display within running the algorithm first time.
This was caused by an unoptimal way of updating the textarea, which is now more optimal.

Changed the methods which converted datasets into one universal method
which decides the appropriate filter to be applied.