Potentially fixed threading issue with a slower PC (laptop)
The laptop we used for testing, that had (Intel i3-4005U, 4GB RAM, Windows 10 Home 64-bit (10.0, Build 17134)) had issues with consecutive clicking of the runAlgorithms button, which would freeze the UI, and generate incorrect results, as well as generate a lot of exceptions.

Added JUnit 5 to pom.xml for implementation of unit tests.